move into own hook
